:recycle: Refactor array select for to inside screen
