+++ src/screens/RegisterScreen/utils/select.js
